Move thread_pool_executor headers to executors module
