+++ libs/execution/CMakeLists.txt
@@ -31,7 +31,6 @@ set(execution_headers
     hpx/execution/executors/polymorphic_executor.hpp
     hpx/execution/executors/rebind_executor.hpp
     hpx/execution/executors/static_chunk_size.hpp
-    hpx/execution/executors/thread_pool_executor.hpp
     hpx/execution/traits/detail/vc/vector_pack_alignment_size.hpp
     hpx/execution/traits/detail/vc/vector_pack_count_bits.hpp
     hpx/execution/traits/detail/vc/vector_pack_load_store.hpp
@@ -68,7 +67,6 @@ set(execution_compat_headers
     hpx/parallel/executors/post_policy_dispatch.hpp
     hpx/parallel/executors/rebind_executor.hpp
     hpx/parallel/executors/static_chunk_size.hpp
-    hpx/parallel/executors/thread_pool_executor.hpp
     hpx/parallel/traits/vector_pack_alignment_size.hpp
     hpx/parallel/traits/vector_pack_count_bits.hpp
     hpx/parallel/traits/vector_pack_load_store.hpp
