+++ libs/executors/include_compatibility/hpx/parallel/executors/thread_pool_executor.hpp
@@ -10,7 +10,7 @@
 #include <hpx/execution/config/defines.hpp>
 #include <hpx/execution.hpp>
 
-#if HPX_EXECUTION_HAVE_DEPRECATION_WARNINGS
+#if HPX_EXECUTORS_HAVE_DEPRECATION_WARNINGS
 #if defined(HPX_MSVC)
 #pragma message(                                                               \
     "The header hpx/parallel/executors/thread_pool_executor.hpp is deprecated, \
