+++ libs/runtime_local/include/hpx/runtime_local/pool_executor.hpp
@@ -11,7 +11,7 @@
 #include <hpx/config.hpp>
 
 #if defined(HPX_HAVE_POOL_EXECUTOR_COMPATIBILITY)
-#include <hpx/execution/executors/thread_pool_executor.hpp>
+#include <hpx/executors/thread_pool_executor.hpp>
 #include <hpx/modules/threadmanager.hpp>
 #include <hpx/runtime_local/runtime_local_fwd.hpp>
